Fixed several recent bugs:
Give an error if there is an unknown file extension.
The @shadow code **must not** use a dummy file extension:
that would cause chaos if later the extension becomes known.

The read code now adds a trailing newline to the old public file if needed.
This seems to be essential: without it sentinels may get appended to the last line.
Chaos would ensure.

This seems like a drastic solution.  It may be possible to do better in the update algorithm...