Completed first draft of leoQtTree.full_redraw. This method draws only visible nodes.
Typically it takes 0.02 sec to draw all visible nodes in LeoPy.leo
when few nodes are shown.

At present the Leo window isn't being updated while Leo is loading files,
but this is proabably a minor problem unrelated to drawing the tree.
So it appears that the qt plugin is a success.  Hurray!