Fixed the find crasher.
The problem was that the leoQtTree code was not being informed that editing of the headline had ended.
Apparently, this was not a problem with the Tk code, but it caused hard crashes in Qt.

The fix was to call a new method, killEditing, in leoTree.initNextText.
This method need not exist: it is called only if hasattr(tree,'killEditing').

All unit tests pass for Qt.  Probably one test still hangs for Tk.