An urgent update to qtGui plugin to eliminate hard crashers. Most of the work involves tree.edit_widget. Most, but not all, calls to tree.edit_widget were eliminated from Leo's core. Some changes required slight changtes to tkGui plugin.
All unit tests pass, with both tk and qt.
However, some important typing tests presently fail with tk.
This is probably only a problem with the unit tests themselves,
but that is not guaranteed.

**Important**: the new code is likely substantially more robust
than the old, especially with the qt plugin enabled.
However, this is an emergency update, and the code has not been
tested extensively in practice.  Use with care.