Added an ugly hack to work around a bad Qt bug. QLineEdit widgets typically generate KeyPressed events, but QLineEdit widgets in QTreeWidgets generate KeyReleased events.
The hack involves comparing the event widget with c.frame.tree.edit_widget(),
which appears to work on XP, but is not guaranteed to work elsewhere.
I'll be running tests on Unbuntu soon.

I'll be discussing this problem with the Qt people.
It *is* fixable, provided they are willing to add a method that changes
how QTreeWidgets work.