Significantly improved traceing in wxLeoTree and baseNativeTreeWidget.
wx.expand/contract item calls w.SelectItem.
For unknown reasons, this is required to draw the tree correctly.

It appears that wx is generating usable events.

However, the screen flashes unacceptably.
In the past I have tried to buffer the screen with no joy.
Perhaps there is a way, this would be a show stopper.