This is the first commit of a truly incremental syntax colorer. All unit tests pass.
Warning: this code is *experimental*.
It's possible that the colorizer can get out of synch.
If so, just select another node to resynch.

Except for changing string delims, this code will typically
recolor only 1-3 lines, regardless of the length of body text.
This can result in a speedup of more than 1000x for large text.