The qt plugin now enables and disables coloring properly. This is extremely complicated to get right, but all appears to work.
In particular, @killcolor for large text does suppress coloring.

Important: Selecting a large text with @killcolor takes the same the
same time as selecting a large text with no QSyntaxHighlighter at all.
This means that the approx 1 sec. time to select a node containing
qtGui.py arises from the single call to w.setAllText.
I'll be verifying this shortly.