Fixed the long-standing "comment" bug.  Hurray!
The problem was a call to match_mark_previous in the Python colorer.
The solution, perhaps not completely correct, is to clear self.prev in recolor.
So this happened not because of Qt at all, but because self.prev was not cleared during incremental coloring.

It's now clear why the bug was "intermittent":
comments had to look like a function call to trigger it!
For example, the bug happened when removing the '#' on this line::

    # g.trace(whatever)
    
All unit tests pass.