Fixed one big colorizer bug and generalized the computation of state.
colorize now does a full recolor if the language has changed.
This kind of logic can not be done in recolor because recolor is called for every line.
It may be possible and desirable to push more of this kind of logic into colorize;
it does not fit easily with recolor.

The state code probably doesn't do any particular good, but it does no harm