Added (presently disabled) prototype of new coloring scheme to qtGui.py. All should be exactly as before.
The new code only defines a restarter for match_span.
This should be the *only* restarted needed for python (!!)

As you can see, defining a restarter is a bit clumsy.
However, even so, the work is finite and routine, so the concept should work.

Important: restarter need to be defined only for pattern matchers that could possibly span lines.