Added "headkey1" and "headkey2" event handlers (hooks). It's not clear whether these will be of much help to any plugin.
It would be possible to define an textChaanged signal handler in
leoQtHeadlineWidget, as shown by the commented-out code in the ctor.

Instead, baseNativeTree.onHeadChanged calls the "headkey1" and "headkey2" hooks.
Note that the hooks pass "s" arguments instead of "ch" arguments.

I can make some changes to these hooks if desired, but I can't change the
information that Qt supplies to the textChanged signals.

All unit tests pass.