fixed support for haskell
Leo as-is does not support Haskell because it uses "--" for the line
comment separator instead of "-- " (note the space).  This causes
problems when attempting to compile the derived files.  (Arguably the
GHC compiler should not be so picky, but it is the way that it is.)

As a fix for this, I have added an entry for "haskell" to the
language deliminators dictionary (as well as to the extensions
dictionaries) which uses the underscore hack to force placement of a
space between the comment marker and the start of the sentinel.