leoSettings.leo: Added accelerators for add-comments and delete-comments command.
Fixed a major, but rare, bug in createMenuEntries:
use the canonical spelling for the actual binding (the stroke).
That is, we set addKey=False in the call to k.shortcutFromSetting only
when computing the "pretty" (and non-functional!) accelerator for the menu.

All unit test pass, but this could be called a major chnage.
Only more testing will determine whether all keystrokes are handled correctly.