Attempted a fix for the Tk undo problem described in bug 498692: Undo deleted node renames current node's headline
This involved a simple, but potentially far-reaching change:
disabling the call to self.endEditLabel in leoTkTree.redraw_now.

All unit tests now pass, and the bug appears to be fixed,
but please use caution.

BTW, the undo/redo methods explicitly call c.bodyWantsFocusNow.
This may be dubious, but it is hardly a major bug.