Added disabled hack to php.py.
To enable the hack, set use_php_hack to True at the start of the file.
This enables some code at the very end of the file.

At present, this does not seem to be satisifactory.  The php test node in test.leo
is not colored properly.

I'm upping this change now so that Gil can play with it.