A mass update revising and simplifying the use of encoding arguments.
- g.toUnicode now uses g.app.defaultEncoding as the default.

- Several methods, including c.setBodyString, leoImport.createOutline,
  c.appendStringToBody, p.setBodyString (etc) no longer take
  any encoding param.

- qtGui.py now uses g.u almost exclusively.

The result of all of this is that it is now easy to determine where
encoding really must be used, namely in read and import methods.
Just search for g.toUnicode(x) where x is non-empty.

All unit tests pass, and pylint is happy, at least on this account.

However, there is no guarantee that the code works in all environments.
Only time will tell.