Several important changes to Leo's read logic to handle clone conflicts:
- at.readStartAll sets at.atAllFlag, initied in initReadIvars.

- at.readEndNode now *does* warn about changed nodes,
provided that at.atAllFlag is set.
The message is: "uncached read node changed <<node>>".

- fc.createSaxVnode now has some more trace,
but these are disabled by default.
    
- v.createOutlineFromCacheList sets atAll at the top level
and passes atAll to itself.  A newly-read cloned node overrides
a previous value only if atAll is False.
    
-  v.createOutlineFromCacheList gives the warning,
"cashed read node changed <<node>>"
if a clone overrides a previous value.
    
- initAfterLoad in qtGui.py no longer calls c.setChanged(False).
This is needed so that updated clones can set the .leo file changed.

- fc.getLeoFile now helps out v.createOutlineFromCacheList, which
can only set the dirty bit for updated vnodes.  The outline does not
exist in a state suitable for p.setAllAncestorAtFileNodesDirty when
v.createOutlineFromCacheList executes.  Therefore, fc.getLeoFile
calls fc.propegatgeDirtyNodes if c.changed is True on exit.
    
- Added leo/test/testCloneUpdate.leo and related files.
This is not as good a test as a unit test would be, but it has been useful.

I created these notes from the node "Clones in @all files now have low priority"
in leoProjects.txt