- Created g.internalError and call it from several places.
- _contentHashFile now calls g.internalError if the content
  arg is unicode.

- Avoid a crash in g.getPythonEncodingFromString by performing a
*temporay* converion of the string/bytes to unicode.

- Changed several unit tests so they work with Python 2x/3x.
  In particular, we can not call g.ue in Python 3.x.
  
- The 'globals' content arg to _contentHashFile in the code that reads
and writes global data now is either leo2k.globals or leo3k.globals.
This allows Leo to switch back and forth between leo2k and leo3k without
clashes in window position.
Of course this string is converted to bytes as well, so that
_contentHashFile does not complain.

All unit test pass with 2.x/3.x on Windows 7.  They will be
run on Ubuntu shortly after this rev is pushed.