Removed trace in qtGui.py.
Added an important new script in unitTest.leo that attempts to import plugins.
On the basis of the ouput from this script, all plugins have been assigned to
one of several lists in the script.  You can double-check these lists easily by
running the script only on items from the list.