Warning: fixed(??) a major bug in the backup logic for .leo files.
- createBackupFile now calls shutil.move(fileName,backupName)
where backupName is the fileName of the closed just returned from mkstemp.
This avoid the call to g.utils_rename, which does too much.

- Removed the extremely dubious call to g.utils_remove(dst) in g.utils_rename.
This duplicates part of the work of shutil.move, and could easily interfere with it.

All unit tests pass, as do several hand tests, but this is a major change.  Use caution.