The start of a complete rewrite of c.goto_countLines.
The code sorta works now: it is much better than the old off-by-a-gazillion code.
Now it's only off by a few.  More work and unit tests are coming.

I'm committing this now so I can start work on improved caching.
This work must be done asap, so it can be tested as much as possible before rc1.