Fixed bug 363406: goto global line doesn't seem to be working in @auto files.
This is similar to rev 2977 of the "until" branch.

QQQ
Rewrote countLinesHelper and created countLinesInChildren.

Heavily revised @test goto-global-line.

The @auto node must be a child of this unit tests so that Alt-6 will work properly.

These counting methods are some of the trickiest I've ever written.
See the invariants in countLinesHelper for the essence of the algorithm:
the trick is that n doesn't change, and that we must keep track of both
skipped_lines and effective_lines.

The unit tests are far more careful than the usual unit tests.

Even for "mister recursion" this code is something special.
QQQ

Removed @test c.gotoLineNumber: it was conflicting with the new, much stronger unit tests.