The start of the read/write code for simplified thin sentinels.
The switches new_read and new_write at the start of leoAtFileCommands
enable the code and show where the new code is.

All unit tests pass with these switches set to False.

Much more work is needed, but this work should be safe enough for the trunk.

All essential parts of the work will be confined to leoAtFile.py.