(1) The code checks for tangleCommands.head_root not none, but nowhere assigns head_root.  To confirm my understanding, I added a number of
assert self.head_root == None

and found that they never triggered during testing.  I speculate that they're
there for future implementation of @root in the headline.

(2)
The testAtRootFile() code was trying to test untangle when @silent or @quiet
were in effect, and untangle is impossible by design.  I stopped doing that.

(3)
Sometime in the past, @doc sections were emitted as comments into the
derived files, and code was present in untangle to ignore - and thus
remove - such comments, at the cost of sometimes losing regular
comments.  Since Leo no longer emits @doc sections into the derived
file, I removed the elimination of those comments from the @code
sections during untangle, thus preserving some interlinear comments.

(4)
Of the unit tests that now fail, they are:
(a) a bug in the test code for multiple @root directives in a single node
(b) a Leo bug or a failure to understand the conditions in which a section
may not have multiple definition parts (sometimes it can)
(c) speculation on how mixing @language and @comment would