We now have a perfectly usable way of creating slides!
1.  Wrote slide names '%03d' format, to create, for example,

    what-is-leo-001
    what-is-leo-002 etc.

This is needed to make the toctree:: :glob: directive properly sort file names.

2. slides.html now contains a top-level toc.  Like this::


    .. toctree::
        :maxdepth: 1

        installation-001
        what-is-leo-001

As shown slides.html must know the file name of the first slide of each slideshow.

3. The first slide of each slideshow must contain a toc for that slideshow.  For example:

.. toctree::
    :hidden:
    :glob:

    installation-*

Note that the :hidden: directive is optional, but it is useful nontheless.

Still to do:

It's a bit off-putting to have one slideshow merge directly into the next.
It would be good to kill (or alter) the "next" link in the last slide of each slideshow.