leoRst.py: created separate method to write the title of each slide.
leoDocs.py: create @button screenshot and got it working.

This script supports @screenshot-tree nodes::

    @screenshot-tree <headline to select>

If present, @button screenshot hoists the first child of the @screenshot-tree node
and takes a picture of *that* tree, *not* the presently selected tree.

Whatever follows @screenshot is taken to be the headline to select.

At present, the user will be responsible for inserting an .. image:: reference into the body text.
That is not odious: the name of the file will not change: it is the gnx of the selected node.