1. Added bullet-proof encoding code to g.pr for Python3k.
The new code should make it impossible for the following to fail:

    sys.stdout.write(s2)

If it does fail, we'll just print a message and attempt no further recovery.
Besides, the previous recovery was bogus.

2. Enabled tracing in writeAllHelper.  This was not the problem, but the unadorned message was confusing.

Summary
======

All unit tests pass with Python 2k/3k on windows. Tests will be run on Linux asap.

This is a preliminary attempt to fix bug 632643:
not opening leo file when there are non-english characters in the path

However, the present tests only deal with external files, not .leo files.