Preliminary work re bug 577047: Invalid @path directory does not warn user
The biggest substantial change is the following pattern in all calls to at.scanDefaultDirectory.

    at.errors = 0
    at.scanDefaultDirectory(...) # Set default_directory
    if at.errors: return

This ensures that errors in the helers will eventually be handled.

However, the write logic may still have to set dirty and orphan bits.

All unit tests pass, which isn't completely trivial, but is not conclusive.
More unit tests for the complex inner helpers is surely needed.