Changed Leo's error recovery logic.  This fixes bug 577047: Invalid @path directory does not warn user.
An change earlier this year, in writeAllAtFileNodesHelper, causes Leo to ignore write errors.

Thus, the *read* logic must be more careful to ensure that nodes are written properly.
To do this, the read logic now sets the orphan bit in the root node when there are errors.
The write logic writes the entire tree to the .leo file if the orphan bit is set.