The first successful test of complete reworking of the @button make-nsi.
I could not bear to contemplate trying to explain the horrors of the old script.
Instead, I replaced it with boilerplate files nsi-xxx.txt.

The point is that only two files, nsi-uninstall-files and nsi-install-files,
will have to be generated from the bzr manifest.
Everything else can, and is, writting in straightforward .nsi style.