Save and restore focus when actually saving: alas, this is not really sufficient.
The problem can only arise when inserting a new node.
This *in itself* changes the outlines, so the autosave will happen later
regardless of the state of the headline.

Alas, it is *not possible* to restore focus properly in headlines: the
QLineEdit widget gets reallocated when the tree is redrawn.

A better solution will be to disable auto save when focus is in the tree.