Eliminated all code that saves and restores the body pane's scrollbar position, using instead w.see and w.seeInsertPoint.
My *guess* is that there were (rare) cases when the w.see clashed with setting the scroll point.
The new code is simpler and can not cause such clashed.  More testing is needed to see if it works well enough.

All the new code is marked with ### 2011/09/30.
All unit tests pass.