An almost complete fix for the scrolling problems encountered at the end of long body text.  Clicks can still cause lesser problems.
We could eliminate mouse-click problems by omitting the call to k.keyboardQuit in the mouse-release handler, but that would
cause mouse-clicks not to reset the key state, and that would be confusing.

The new code is much better than before, and I don't know how to improve it further.

The Aha is that trying to save & restore scrollbar state does not work: don't ask me why.
Instead, the fix is to call either w.seeInsertPoint (for high-level w) or w.ensureCursorVisible (for low-level w).

All unit tests pass, which is significant: some seemingly innocent changes cause many tests to fail.