An important improvement to unit tests: hidden unit tests.
A hidden unit test is a node that is *not* a section-definition node and whose
body text contains one of the following guards at the start of a line::
    
    if g.unitTesting
    if g.app.unitTesting:
    if g.inScript:
    if g.app.inScript:
        
Any one of the guards will prevent the unit testing code from being executing
when importing a file.

The typical usage pattern is::
    
    class myClass:
        ...
        
    if g.unitTesting:
        x = myClass()
        x.test()

For leoPy.leo, I redefined alt-4 to mean run-all-unit-tests-locally, which is just what is wanted.