Fix bug 944555: Ctrl-left-click URL handling not as sophisticated as @url URL handling
A number of fixes related to unifying URL handling throughout Leo:

- leoFrame.py:

OnIconDoubleClick now gets the url using g.getUrlFromNode.

- leoGlobals.py:

g.isValidUrl now returns True only if the scheme starts with one of:
    
 'file','ftp','gopher','hdl','http','https','imap',
 'mailto','mms','news','nntp','prospero','rsync','rtsp','rtspu',
 'sftp','shttp','sip','sips','snews','svn','svn+ssh','telnet','wais'

This is exactly the list of schemes supported by urlparse,
but strangely, urlparse may return as scheme that does not start with one of the above.

This change is important, otherwise a line like:
see:  <some url>
will cause the url to be opened when double-clicking the node,
which is not likely to be what is wanted.

- bookmarks.py:

onDClick1 now gets the url using g.getUrlFromNode.

- contextmenu.py:

Adds an "Open URL" rclick item if the headline looks like an url.
Naturally, the code calls g.getUrlFromNode to make this decision.

With these changes, it appears that all url handling throughout Leo
and its plugins uses the same code.