Completed the first draft of the conversion to new-style IPython Api.
Notes:

1. At present, only one version (legacy or new-api) of IPython should be on the path.
This probably could be changed if leoIPython.py actually attempted to import only
one version, but at present it attempts to load the new-api first, and then the legacy api.

This won't work if both are present in the path, because so new-api imports succeed,
and that apparently messes up the imports of the legacy version.

2. start_new_api now just calls::

    ipapp.launch_new_instance()
    
This prints the standard IPython signon.  Later, update_commander sets the ip ivar.

3. I changed the push-to-ipython command so that it always calls the
startIPython code.  This ensures that the proper commander is always in effect.

4. The push_to_ipython command now uses the same code regardless of IPython version.

5. Very few changes we needed to support the new api:

- A few new imports in << attempt to import new version of IPython >>.

- A few api-specific changes in get_history, push_ipython_script and
  mb_completer.  Only push_ipython_script has been tested.

  It might be possible to get a completer from the present shell (self.ip),
  bug I haven't done that.

In short, the code is now ready for real testing.