Fixed another import fail in an "interesting" way: the import code no longer inserts @verbatim.  This means a later write of the imported will fail.  This is correct!
Indeed, the failed write is the only way to alert the user that the code must be revised by hand.

Note that another import fail, involving a leading '@' on a line in a docstring, must also be fixed by hand.
In lib2to3/pgen/grammar.py the *only* possible fix is to enclose the entire docstring at the end of the file
by @raw and @end_raw.

All unit tests pass, but no new tests have been added so far.