Fixed bug 1208659 leo parsed the wrong line number of html file.
The compare failure was the result of using baseScannerClass.startsString.

This is wrong: single quotes do *not* start strings in xml or html.



Added better traces when the comparison fails.



Another change: if the @auto import fails, the file is imported as usual
(with an @ignore node). 
That is, no @edit node is created. This disables a
dubious feature.



Still to do: improve the misleading error message itself.