Completed the fix of bug 1208659 leo parsed the wrong line number of html file. https://bugs.launchpad.net/leo-editor/+bug/1208659
The proximate cause of the erroneous dump was a blunder in setting the fail_n2 in compareTokens.
It was set to the same value as fail_n1! No wonder the lines match.

Alas, reportMismatch can not detect tokenizing errors:
reportMismatch only has the imported lists of *lines* for the error report, not the tokens.
In other words, the only way to detect tokenizing errors is with unit tests.

This concludes work on this bug.