The grand changeover to fully-parameterized style-sheets. Almost all settings in @data qt-gui-plugin-style-sheet in leoSettings.leo are now defined in terms of @values. Traditional Leo settings nodes (@strings, @color, @font, etc.) define these @values.
Important: Almost all existing Leo users are going to have to adjust settings.  This can't be helped and I apologize in advance.
The place to start is with the node called "Basic Fonts (change these first)"  This should be enough to get things looking pretty good for you.

Leo build: 20140916102633