Fixed bug 36: @auto file not found produces a script error.
The report seems misleading: it says this happens when Leo tries to load an @auto file that is in the outline but which's file was erased outside Leo, but it worked for me.

I did add a guard in g.recureiveUNLFind to ensure that the IndexError shown in the bug report can never happen.

Leo build: 20141026060621