Reverted to last night's code, with a few minor changes. The new full-scan code is too slow for large text. It may be useful in the future, but I doubt it.  The new code is in that attic.
All tests pass, as does a test script.
However, back scanning still fails on this snippet:

    {
        a + "//" + b
    }

The solution (reasonable for the match-bracket command) is to assume that
the bracket being match is inside neither a string nor a comment nor a regex.

If we ever use the code for importing, forward scans must be used exclusively.

Leo build: 20160121172053