Added #272: optionally delay syntax coloring. Investigated #273: Qt Bug involving large text nodes.
These two items are related. The motivation is to avoid pauses when skipping over large text.
The present code is likely the best that can be done.

This is a straightforward change.  It's benign, but really does not solve the problem.

The ultimate source of the delay when selecting large text is *not* syntax coloring,
but instead the call to self.old_w.setPlainText(self.p.b) in btw.create_widgets.

Experiments show loading "dummy" text in btw.create_widgets can lead to data loss.
There are too many paths through the selection logic to attemp to "recover"
from the failure to load the body widget initially.

Obviously, data loss is a much bigger problem than slow load times.
I am going to declare this issue not to be solvable.