Created a new i.cut_stack method, following the python.cut_stack method.  All tests pass.
The only difference is that i.cut_stack does *not* pop the stack when new_state.level() == prev_state.level().

The python and coffeescript versions of cut_stack are identical.

There are several ways of handling these differences, but the simplest is for python and coffeescript
to override i.cut_stack as they do now.  This will allow them to use i.v2_gen_lines, the final unification.

Leo build: