A major milestone: Replaced py_i.v2_gen_lines by a *changed* i.v2_gen_lines.
Presumably, using i.v2_gen_lines instead of coffee_i.v2_gen_lines will now be straightforward.

There were two tricky distinctions involved:

1. Testing levels

i.v2_gen_lines now tests:

    if new_state.level() >= top.state.level():

instead of:

    if self.continues_block(line, new_state, prev_state):

In other words, the correct test is *always* between new_state and the top
state on the stack, *not* the previous state.

I can live with the change because all unit tests pass.  It's the kind of detail
that is impossible to understand for long.  It had better be correct...

As a happy side effect, the continues_block methods are no more. That's
good, because this was a hard concept to understand.

**Important**: py_i.v2_gen_lines is gone, but the proper operation of the
python importer depends on overrides of i.v2_gen_lines methods.

2. There is a subtle, crucial, difference between i.cut_stack and both
   py_i.cut_stack and coffee_i.cut_stack. The python and coffeescript
   versions of cut_stack must *not* pop the stack one more time when
   stack.top.state == new_state. This is an *essential* feature of
   indentation-based languages.

Leo build: