Suppress a warning in at.scanAllDirectives when using the markdown importer.
Markdown is a special case because it has no comment delims.

It was surprisingly hard to find a workaround.
The method chosen isn't great, but it's better than more global hacks.

There may be a subtle bug in at.scanAllDirectives,
but I definitely don't want to change it just for this one special case.

Leo build: