Lots of experimental work, enabled by the NEW ivar in QScintillaColorizer.
When NEW is True:

- qsc.highlighter = PythonHighlighter(c)

- qsc.colorizer calls jedit.recolor "manually", for each line.
  Of course this is slow, but it proves that the jEdit colorizers *could* be used.

At present, no format changes are applied because QScintillaWrapper.tag_configure is a do-nothing.

The more important question is how to recolor only those lines that need to be recolored.
The attic contains many additions to the present PythonHighlighter class.
This is ancient history, but it is possible, in theory, to colorize until no state changes are possible.

None of this work is a big improvement to existing scintilla lexers.
As I think about this, it's probably more important just to support all existing scintilla lexers.

Leo build: 20170128125018