Fixed #387: New file is marked as changed. https://github.com/leo-editor/leo-editor/issues/387
The problem:

abbrev.init_tree_abbrev changes the outline's changed indicator (*) during startup.
Simply reversing this in c.new causes the * to flip back and forth.
That's not good enough: we can't have this kind of flash.

The changes were a bit tricky:

- init_tree_abbrev calls c.pasteOutline(...,redrawFlag=False)
- pasteOutline calls c.setChanged(True, redrawFlag=redrawFlag)
- c.new calls c.setChanged(False)
- Added redrawFlag keyword arg to c.setChanged.
  When this is false, c.setChanged does not alter the outline's changed indicator.

Suppressing the change indicator is a bit dangerous.
It would be very confusing if the indicator gets out of sync with reality.
But the fix should actually be safe enough, because only startup code is affected.

Leo build: 20170131045907

Leo build: 20170131052323