All unittest finally pass on my machine
Apparently reason for 6 failing tests in leoRst set of tests was that option call_docutils was not set.
In the Leo documentation it is stated that default is True, but perhaps it was overriden in myLeoSettings.leo
and that caused docutils to never get called. That is why those six tests used to fail.

Solution was to add block of at-rst-option in those six tests.