Fixes #457
Search now remembers expanded state of all nodes when search starts, and
restores it when search has been exhausted.

In current implementation every clone of a node is treated the same way. It gets
either expanded or contracted regardless of where it is in outline tree. Leo allows
clones to be separately expanded/contracted on different positions, however
search commands do not support this feature (yet).