Added command for toggle split direction just for two adjacent widgets.
New command is in context menu of splitter handles. If splitter handle
is vertical then it takes left and right component and repositiones them
in vertical orientation as above and below. If handle is horizontal then
it repositiones them to lay left and right of handle. Preserving ratio
between those two components and the rest of layout.

Menu label is 'Turn split/dir. just this'.