Solves problem with vierendered and cache python2/3 versions.
It appeared that some values in stored layout were in fact
PyQt enumeration instances although they can be represented
with integers. This made unpickling impossible because of
differences in PyQt versions. Solution was to dump layout
values in json and to recreate them from json, and then to
store them in db.

Leo build: 20170818100409