Fixes #540
Added three new commands:

   1. `set-reference-file` which selects which reference Leo file will be this outline connected to
   2. `read-ref-file` which reads public part of this outline from reference Leo file
   3. `update-ref-file` which writes public part of this outline to reference Leo file

These commands rely on convention of having one top-level node with the headline set to
'---begin-private-area---'. This node has in the first line of body path to reference
Leo document. Everything after this node is considered as private part of the outline
and it won't be writen in reference file. Every node before this node is considered to
be public part of the outline. This public part will be written to reference file using
command `update-ref-file`, and this public part will be completely reconstructed from
the reference file using command `read-ref-file`.

Leo build: 20170831135210