The run_travis_unit_tests.py script runs all unit tests and does sys.exit(0) if all is well, else sys.exit(1)
All of Leo's unit tests continue to pass, when run the normal way.
An important change to Leo's core: The leoQt module now *always* allows qt imorts, even from the bridge.

To do: the leoBridge module does not init the externalFilesController module properly.
Several unit tests were disabled when g.in_bridge is True to accommodate this.
This should be fixed.  I'll create an issue for it.